refactor: create fetchScema with createAsyncThunk
