+++ src/redux/slices/GraphQLSlice.ts
@@ -5,6 +5,8 @@ import {
   UnknownAction,
 } from '@reduxjs/toolkit';
 import { BASE_API_LINK, QUERY_TEMPLATE } from '../../shared/constants';
+import { parseSchema } from '../../utils/parseSchema';
+import { getIntrospectionQuery } from 'graphql';
 
 type InitialState = {
   apiLink: string;
@@ -33,6 +35,10 @@ type FetchJSONParams = {
   query: string;
   variables: string;
   headers: string;
+};
+
+type FetchSchemaParams = {
+  url: string;
 };
 
 export const fetchJSON = createAsyncThunk<
@@ -77,6 +83,36 @@ export const fetchJSON = createAsyncThunk<
   }
 );
 
+export const fetchSchema = createAsyncThunk<
+  string,
+  FetchSchemaParams,
+  { rejectValue: string }
+>('GraphQLSlice/fetchSchema', async ({ url }, { rejectWithValue }) => {
+  try {
+    const response = await fetch(url, {
+      method: 'POST',
+      headers: {
+        'Content-type': 'application/json',
+      },
+      body: JSON.stringify({
+        query: getIntrospectionQuery(),
+      }),
+    });
+    const data = await response.json();
+    if (!response.ok) {
+      const message = (await response.json()).errors[0].message as string;
+      return rejectWithValue(`Schema: ${message}`);
+    }
+    const schema = parseSchema(data.data.__schema.types);
+    return JSON.stringify(schema, null, 4);
+  } catch (error) {
+    if (error instanceof Error) {
+      return rejectWithValue(`Schema: ${error.message}`);
+    }
+    return rejectWithValue('`Schema: something went wrong!');
+  }
+});
+
 const isError = (action: UnknownAction) => {
   return action.type.endsWith('rejected');
 };
@@ -112,6 +148,9 @@ const GraphQLSlice = createSlice({
     builder
       .addCase(fetchJSON.fulfilled, (state, action) => {
         state.jsonViewer = action.payload;
+      })
+      .addCase(fetchSchema.fulfilled, (state, action) => {
+        state.documentation = action.payload;
       })
       .addMatcher(isError, (state, action: PayloadAction<string>) => {
         state.error = action.payload;
